requirements.txt
streamlit
python-dotenv
llama-parse
langchain
langchain-core
langchain-community
langchain-google-genai
langchain-huggingface
langchain-chroma
chromadb
python-dotenv
